Hooks update.
This is a big update.
At first I updated HTTP hooks, so they
proxy status code and response contents in case of error.

Then I fixed TUSD format it did not match real tusd format.
Also I updated docs and examples for all formats. Related to #90.

And in the end I added `--behind-proxy` parameter to
be able to get real IP's of clients.

Signed-off-by: Pavel Kirilin <win10@list.ru>